Implement multi-agent orchestrator, background tasks, schedule tools, and fix memory search
Orchestrator crate with supervisor/delegation pattern: DelegationTool for
single-agent delegation and ParallelDelegationTool for concurrent multi-agent
work (up to 5). Uses LazyOrchestrator (OnceCell) to resolve init ordering
between tool registry and orchestrator. Agent definitions loaded from YAML
in agents/ directory with per-agent model, system prompt, and tool config.

Background task system: spawn long-running agent work that persists to
background_tasks table, runs in tokio tasks, and notifies via broadcast
channel. SSE endpoint streams task notifications to gateway, which forwards
to WebSocket clients per-user for real-time push.

Schedule management tools (schedule_list, schedule_create, schedule_update,
schedule_delete) registered in the agent's tool registry so the agent can
create and manage cron jobs from chat — fixes the bug where the agent said
it had no scheduling capabilities.

Fix memory search NUMERIC type mismatch: RRF score computation produces
NUMERIC in Postgres but Rust expected f64. Added ::float8 casts.

Full stack: core routes, gateway routes, frontend API types, Tasks admin
tab with spawn/view/cancel, notification store with sidebar badge.

+++ core/crates/memory/src/postgres.rs
@@ -122,13 +122,13 @@ impl MemoryBackend for PostgresMemory {
                         COALESCE(f.category, v.category) AS category,
                         COALESCE(f.confidence, v.confidence) AS confidence,
                         COALESCE(f.created_at, v.created_at) AS created_at,
-                        COALESCE(1.0 / (60 + ROW_NUMBER() OVER (ORDER BY f.fts_score DESC NULLS LAST)), 0) AS fts_rrf,
-                        COALESCE(1.0 / (60 + ROW_NUMBER() OVER (ORDER BY v.vec_score DESC NULLS LAST)), 0) AS vec_rrf
+                        COALESCE((1.0 / (60 + ROW_NUMBER() OVER (ORDER BY f.fts_score DESC NULLS LAST)))::float8, 0) AS fts_rrf,
+                        COALESCE((1.0 / (60 + ROW_NUMBER() OVER (ORDER BY v.vec_score DESC NULLS LAST)))::float8, 0) AS vec_rrf
                     FROM fts_results f
                     FULL OUTER JOIN vector_results v ON f.id = v.id
                 )
                 SELECT id, content, scope, category, confidence, created_at,
-                       (fts_rrf + vec_rrf) AS combined_score
+                       (fts_rrf + vec_rrf)::float8 AS combined_score
                 FROM combined
                 ORDER BY combined_score DESC
                 LIMIT $5
